Distributed computing infrasrtucture using smartphones
Every night, many smartphones are plugged into a power source for recharging the battery. Given the increasing computing capabilities of smartphones, these idle phones constitute a size able computing infrastructure. Therefore, for an enterprise which supplies its employees with smartphones, it is arguable that a computing infrastructure that leverages idle smartphones being charged overnight is an energy - efficient and cost-effective alternative to running certain tasks on traditional servers.